workerino

+++ program/benches/my_benchmark.rs
@@ -1,6 +1,6 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 
-fn fibonacci(n: u64) -> u64 {
+fn fibonacci(n: u128) -> u128 {
     match n {
         0 => 1,
         1 => 1,
